test(research): split PDF table layout coverage
